+++ src/components/domain/Directory/molecules/DirectoryListItem/index.tsx
@@ -10,7 +10,7 @@ import { BootstrapBreakpoints } from '~/interfaces/variables';
 
 import { useDirectoryForDelete, useParentDirectoryForCreateDirectory, useDirectoryForRename } from '~/stores/modal';
 
-import { IconButton } from '~/components/Icons/IconButton';
+import { IconButton } from '~/components/base/molecules/IconButton';
 import { Icon } from '~/components/base/atoms/Icon';
 
 type Props = {
